refactor(cli): move help module under commands
