+++ crates/tsz-cli/src/commands/help.rs
